backend/requirements.txt
# LiveKit Agents Framework (v1.0+)
livekit-agents[openai,silero,sarvam,turn-detector]~=1.0

# LiveKit API for token generation
livekit-api

# Configuration
pydantic-settings

# FastAPI for token API server
fastapi
uvicorn

# Environment management
python-dotenv
